+++ Changelog.md
@@ -6,6 +6,11 @@
   (e.g., `os(multics)`) were crashing cabal-to-dhall. They now work as
   expected.
 
+* Remove orphan `Dhall.Core.Inject` instances for `[Char]` and
+  `CompilerFlavor`.
+
+* `dhall-to-cabal` and `cabal-to-dhall` now respond to `--version`.
+
 ## 1.1.0.0 -- 2018-06-03
 
 ### Breaking Changes